Fixed indexing error

+++ code/bestFit.py
@@ -171,7 +171,7 @@ def write_atmfile(atmfile, molfit, rad, T_line, allParams, date_dir):
 
     # recognize which columns to take from the atmospheric file
     headers = lines[start-1].split()
-    columns = np.zeros(len(molfit))
+    columns = np.zeros(len(molfit), dtype=int)
     for i in np.arange(len(molfit)):
         for j in np.arange(len(headers)):
             if molfit[i] == headers[j]:
